Change ref. type to List instead of ArrayList
the reason why we are using ArrayList and not LinkedList is because
we are iterating and adding elements to the end  of the list.
Both LinkedList and ArrayList perform this operation for time O(1)
(in the unfavorable case ArrayList will spend O(n) on this operation -
if we exceeded the size of inner array, ArrayList will perform
array copying - but it is completely another story...)
But. When we look into source code, we can see that ArrayList add
operation just add element to the array. And LinkedList allocates new
object - its inner node.